requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
honeybadger==0.19.0
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
celery==5.3.6
channels~=4.0.0
celery~=5.3.6
